+++ simply-synonyms/content/content.js
@@ -35,6 +35,11 @@ function resetPopup() {
   document.getElementById('ssyn-content').style.display = 'none'
   document.getElementById('ssyn-results-text').innerHTML = ''
   document.getElementById('ssyn-synonyms').innerHTML = ''
+  document.getElementById('ssyn-popup').style.height = ''
+}
+
+function adjustPopupPosition() {
+
 }
 
 function addExtension() {
@@ -52,13 +57,19 @@ function addExtension() {
     getSynonyms(text)
       .then((response) => {
         if (response.synonyms) {
-          const flattenedSynonyms = response.synonyms.flat()
           const synonymsDiv = document.getElementById('ssyn-synonyms')
-          synonymsDiv.innerHTML = ''
-          for (const syn of flattenedSynonyms) {
-            const synEl = document.createElement('div')
-            synEl.innerText = syn
-            synonymsDiv.appendChild(synEl)
+          for (const [i, def] of Object.entries(response.shortdefs)) {
+            // Create definition labels
+            const defEl = document.createElement('div')
+            defEl.classList.add('ssyn-shortdef')
+            defEl.innerText = def.toString()
+            synonymsDiv.appendChild(defEl)
+            // Print out synonyms for each definition
+            for (const syn of response.synonyms[i]) {
+              const synEl = document.createElement('span')
+              synEl.innerText = syn
+              synonymsDiv.appendChild(synEl)
+            }
           }
         } else {
           document.getElementById('ssyn-results-text').innerText = `Unable to find synonyms for "${text}"`
@@ -66,6 +77,22 @@ function addExtension() {
 
         document.getElementById('ssyn-loading').style.display = 'none'
         document.getElementById('ssyn-content').style.display = 'block'
+
+        // Adjust height & position
+        if (window.innerHeight - 20 - popup.offsetTop < popup.offsetHeight) {
+          // Popup is overflowing on bottom of page
+          popup.style.top = `${(window.innerHeight  - 20 - popup.offsetHeight)}px`
+          popup.style.left = `${popup.offsetLeft + 50}px`
+        }
+        if (window.innerWidth - 20 - popup.offsetLeft < popup.offsetWidth) {
+          // Popup is overflowing on right side of page
+          popup.style.left = `${popup.offsetLeft - popup.offsetWidth - 20}px`
+        }
+        if (popup.offsetHeight > window.innerHeight - 40) {
+          // popup is too tall to fit on page, enable scrolling
+          popup.style.height = `${window.innerHeight - 40}px`
+          popup.style.top = '20px'
+        }
       })
   })