+++ changelog.md
@@ -1,4 +1,8 @@
-Simply Synonyms Change Log
+## Simply Synonyms Change Log
+
+#### V0.1.1
+ + Show short definitions above synonyms
+ + Automatically adjust popup positioning after content is added so that it doesn't overflow off side or bottom of page
 
 #### V0.1.0
  + Basic synonym popup with a list of synonyms and a close button